# packages/contracts/openapi.yaml
...
openapi: 3.1.0
info:
  title: Crewquarters Control API
  description: 'Control plane for Crewquarters. JSON is camelCase. Errors use `{error: {code, message,
    requestId, details}}`. State-changing requests need the session cookie, an allowed Origin, and the
    `X-CSRF-Token` header. Mutating requests accept an `Idempotency-Key` header.'
  version: 0.1.0
paths:
  /api/v1/bootstrap:
    post:
      tags:
      - auth
      summary: Create the device owner with the one-time bootstrap token
      operationId: bootstrap_api_v1_bootstrap_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BootstrapIn'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionOut'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/sessions:
    post:
      tags:
      - auth
      summary: Sign in
      operationId: login_api_v1_sessions_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginIn'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionOut'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/runs/{run_id}:
    get:
      tags:
      - runs
      summary: Get a run
      operationId: get_run_api_v1_runs__run_id__get
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Run Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunOut'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
...
  /api/v1/runs/{run_id}/events:
    get:
      tags:
      - runs
      summary: Stream run events (SSE)
      operationId: stream_run_events_api_v1_runs__run_id__events_get
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Run Id
      - name: after
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
            minimum: 0
          - type: 'null'
          description: Resume after this sequence (alternative to Last-Event-ID).
          title: After
        description: Resume after this sequence (alternative to Last-Event-ID).
      - name: Last-Event-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Last-Event-Id
      responses:
        '200':
          description: Server-sent events. Each event has id=sequence, event=<type>, data=RunEvent JSON.
            The stream ends with an 'end' event when the run reaches a terminal state. Reconnect with
            Last-Event-ID to resume without duplicates.
          content:
            application/json:
              schema: {}
            text/event-stream:
              schema:
                type: string
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
...
components:
  schemas:
...
    BootstrapIn:
      properties:
        token:
          type: string
          maxLength: 256
          minLength: 16
          title: Token
          description: One-time bootstrap token.
        username:
          type: string
          maxLength: 64
          minLength: 3
          pattern: ^[A-Za-z0-9_.@-]+$
          title: Username
        password:
          type: string
          maxLength: 1024
          minLength: 12
          title: Password
        email:
          anyOf:
          - type: string
            maxLength: 254
          - type: 'null'
          title: Email
      type: object
      required:
      - token
      - username
      - password
      title: BootstrapIn
...
    ErrorDetail:
      properties:
        code:
          type: string
          title: Code
          examples:
          - MODEL_CAPACITY_EXCEEDED
        message:
          type: string
          title: Message
        requestId:
          anyOf:
          - type: string
          - type: 'null'
          title: Requestid
        details:
          additionalProperties: true
          type: object
          title: Details
      type: object
      required:
      - code
      - message
      title: ErrorDetail
    ErrorResponse:
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
      type: object
      required:
      - error
      title: ErrorResponse
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    LoginIn:
      properties:
        username:
          type: string
          maxLength: 254
          minLength: 1
          title: Username
        password:
          type: string
          maxLength: 1024
          minLength: 1
          title: Password
      type: object
      required:
      - username
      - password
      title: LoginIn
...
    RunOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        installationId:
          type: string
          format: uuid
          title: Installationid
        agentId:
          type: string
          title: Agentid
        agentName:
          type: string
          title: Agentname
        agentVersion:
          type: string
          title: Agentversion
        trigger:
          type: string
          enum:
          - manual
          - schedule
          title: Trigger
        scheduleId:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Scheduleid
        scheduledFor:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Scheduledfor
        state:
          type: string
          enum:
          - QUEUED
          - PREPARING
          - LOADING_MODEL
          - RUNNING
          - WAITING_INPUT
          - CANCELLING
          - SUCCEEDED
          - FAILED
          - CANCELLED
          - INTERRUPTED
          title: State
        currentAttempt:
          type: integer
          title: Currentattempt
        result:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Result
        error:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Error
        retryable:
          type: boolean
          title: Retryable
        cancelRequested:
          type: boolean
          title: Cancelrequested
        acknowledgedAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Acknowledgedat
        activeSecondsUsed:
          type: number
          title: Activesecondsused
        inputWaitSecondsUsed:
          type: number
          title: Inputwaitsecondsused
        activeTimeoutSeconds:
          type: integer
          title: Activetimeoutseconds
        maxInputWaitSeconds:
          type: integer
          title: Maxinputwaitseconds
        usesCloud:
          type: boolean
          title: Usescloud
        pendingInputCount:
          type: integer
          title: Pendinginputcount
        createdAt:
          type: string
          format: date-time
          title: Createdat
        startedAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Startedat
        finishedAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Finishedat
        updatedAt:
          type: string
          format: date-time
          title: Updatedat
      type: object
      required:
      - id
      - installationId
      - agentId
      - agentName
      - agentVersion
      - trigger
      - scheduleId
      - scheduledFor
      - state
      - currentAttempt
      - result
      - error
      - retryable
      - cancelRequested
      - acknowledgedAt
      - activeSecondsUsed
      - inputWaitSecondsUsed
      - activeTimeoutSeconds
      - maxInputWaitSeconds
      - usesCloud
      - pendingInputCount
      - createdAt
      - startedAt
      - finishedAt
      - updatedAt
      title: RunOut
...
    SessionOut:
      properties:
        user:
          $ref: '#/components/schemas/UserOut'
        csrfToken:
          type: string
          title: Csrftoken
          description: Send as the X-CSRF-Token header on state-changing requests.
        expiresAt:
          type: string
          format: date-time
          title: Expiresat
        idleExpiresAt:
          type: string
          format: date-time
          title: Idleexpiresat
      type: object
      required:
      - user
      - csrfToken
      - expiresAt
      - idleExpiresAt
      title: SessionOut
...
    UserOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        username:
          type: string
          title: Username
        email:
          anyOf:
          - type: string
          - type: 'null'
          title: Email
        role:
          type: string
          enum:
          - owner
          - member
          title: Role
        createdAt:
          type: string
          format: date-time
          title: Createdat
      type: object
      required:
      - id
      - username
      - email
      - role
      - createdAt
      title: UserOut
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: cq_session
    csrf:
      type: apiKey
      in: header
      name: X-CSRF-Token
    service:
      type: http
      scheme: bearer
      description: Internal service credential
servers:
- url: /
  description: Same origin as the web UI